Add CLI options parser
